feat: allow overriding install path
This provides a way to override any configuration on the installer and
install to a location specified by a new environment variable. It provides
a more predictable way to insist that it gets installed to a specific place,
regardless of how the installer was built.

Possible fix for axodotdev/axoupdater#32, where we'd like to be able to
install a new version to exactly the same path as the previous version.